pyproject.toml
[project]
name = "home-alog"
version = "0.1.0"
description = ""
authors = [
    {name = "scoobydrew",email = "drewcsmith3@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "neo4j>=5.15.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
